db/requirements.txt
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
sqlalchemy>=1.4
alembic>=1.7
